requirement.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
loguru==0.7.2
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
win32-setctime==1.1.0
